Refactor layout and modularize PMO pages
