+++ pages/_app.js
@@ -1,7 +1,5 @@
 import Head from 'next/head';
 import '../styles/globals.css';
-import Header from '../components/Header';
-import Footer from '../components/Footer';
 
 export default function App({ Component, pageProps }) {
   return (
@@ -15,9 +13,7 @@ export default function App({ Component, pageProps }) {
           rel="stylesheet"
         />
       </Head>
-      <Header />
       <Component {...pageProps} />
-      <Footer />
     </>
   );
 }
