--- pages/g0.js
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-
-export default function G0Page() {
-  return (
-    <>
-      <Head>
-        <title>G0 - Fluxo PMO Educacross</title>
-        <meta name="description" content="Resumo da fase G0 do fluxo do PMO Educacross." />
-      </Head>
-      <header className="page-header-minimal">
-        <div className="page-header-minimal__inner">
-          <h1>G0 — Intake &amp; Triage</h1>
-        </div>
-      </header>
-      <main className="simple-page">
-        <p>Conteúdo detalhado da fase G0 será adicionado em breve.</p>
-      </main>
-    </>
-  );
-}
