--- pages/g1.js
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-
-export default function G1Page() {
-  return (
-    <>
-      <Head>
-        <title>G1 - Fluxo PMO Educacross</title>
-        <meta name="description" content="Resumo da fase G1 do fluxo do PMO Educacross." />
-      </Head>
-      <header className="page-header-minimal">
-        <div className="page-header-minimal__inner">
-          <h1>G1 — Descoberta &amp; Iniciação</h1>
-        </div>
-      </header>
-      <main className="simple-page">
-        <p>Conteúdo detalhado da fase G1 será adicionado em breve.</p>
-      </main>
-    </>
-  );
-}
