--- pages/g2.js
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-
-export default function G2Page() {
-  return (
-    <>
-      <Head>
-        <title>G2 - Fluxo PMO Educacross</title>
-        <meta name="description" content="Resumo da fase G2 do fluxo do PMO Educacross." />
-      </Head>
-      <header className="page-header-minimal">
-        <div className="page-header-minimal__inner">
-          <h1>G2 — Planejamento Detalhado</h1>
-        </div>
-      </header>
-      <main className="simple-page">
-        <p>Conteúdo detalhado da fase G2 será adicionado em breve.</p>
-      </main>
-    </>
-  );
-}
