--- pages/g3.js
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-
-export default function G3Page() {
-  return (
-    <>
-      <Head>
-        <title>G3 - Fluxo PMO Educacross</title>
-        <meta name="description" content="Resumo da fase G3 do fluxo do PMO Educacross." />
-      </Head>
-      <header className="page-header-minimal">
-        <div className="page-header-minimal__inner">
-          <h1>G3 — Execução &amp; Monitoramento</h1>
-        </div>
-      </header>
-      <main className="simple-page">
-        <p>Conteúdo detalhado da fase G3 será adicionado em breve.</p>
-      </main>
-    </>
-  );
-}
