--- pages/g4.js
@@ -1,20 +1,0 @@
-import Head from 'next/head';
-
-export default function G4Page() {
-  return (
-    <>
-      <Head>
-        <title>G4 - Fluxo PMO Educacross</title>
-        <meta name="description" content="Resumo da fase G4 do fluxo do PMO Educacross." />
-      </Head>
-      <header className="page-header-minimal">
-        <div className="page-header-minimal__inner">
-          <h1>G4 — Lançamento &amp; Estabilização</h1>
-        </div>
-      </header>
-      <main className="simple-page">
-        <p>Conteúdo detalhado da fase G4 será adicionado em breve.</p>
-      </main>
-    </>
-  );
-}
