+++ pages/instrumentos/g0.js
@@ -1,20 +1,23 @@
-import Head from 'next/head';
+import Layout from '../../src/components/layout/Layout';
 
 export default function InstrumentoG0Page() {
+  const hero = (
+    <header className="page-header-minimal">
+      <div className="page-header-minimal__inner">
+        <h1>Instrumentos — G0</h1>
+      </div>
+    </header>
+  );
+
   return (
-    <>
-      <Head>
-        <title>Instrumentos G0 - PMO Educacross</title>
-        <meta name="description" content="Instrumentos da fase G0 do PMO Educacross." />
-      </Head>
-      <header className="page-header-minimal">
-        <div className="page-header-minimal__inner">
-          <h1>Instrumentos — G0</h1>
-        </div>
-      </header>
-      <main className="simple-page">
+    <Layout
+      title="Instrumentos G0 - PMO Educacross"
+      description="Instrumentos da fase G0 do PMO Educacross."
+      hero={hero}
+    >
+      <section className="simple-page">
         <p>Os instrumentos específicos para a fase G0 serão publicados em breve.</p>
-      </main>
-    </>
+      </section>
+    </Layout>
   );
 }
