+++ src/components/flow/FlowSection.jsx
@@ -1,0 +1,16 @@
+import SectionBadge from '../common/SectionBadge';
+
+/**
+ * Seção reutilizável do fluxo, com selo numérico e conteúdo personalizado.
+ */
+export default function FlowSection({ badge, children, id, title }) {
+  return (
+    <section className="flow-section" id={id}>
+      <h2>
+        <SectionBadge value={badge} />
+        {title}
+      </h2>
+      {children}
+    </section>
+  );
+}
