+++ src/components/flow/flowSectionsData.js
@@ -1,0 +1,484 @@
+const flowSections = [
+  {
+    id: 'objetivo-principios',
+    badge: '0',
+    label: 'Objetivo',
+    title: 'Objetivo e Princípios',
+    content: (
+      <>
+        <p>
+          Criar um fluxo único, claro e mensurável de ponta a ponta (da entrada à operação e encerramento) para todos os
+          projetos Educacross, reduzindo custos operacionais, aumentando previsibilidade de entregas e suportando expansão.
+        </p>
+        <h3>Princípios</h3>
+        <ul>
+          <li>
+            <strong>Uma Fonte de Verdade:</strong> Bitrix24 como orquestrador de tarefas/fluxos; dashboards (Looker/Metabase) para
+            visão executiva.
+          </li>
+          <li>
+            <strong>
+              Governança com{' '}
+              <a
+                href="https://www.notion.so/O-que-s-o-Stage-Gates-277e9f3bb81e80589759c9043334eed4?pvs=21"
+                target="_blank"
+                rel="noopener noreferrer"
+              >
+                Stage Gates
+              </a>
+              :
+            </strong>{' '}
+            decisões em pontos de controle (Gates G0→G4) para foco e eficiência.
+          </li>
+          <li>
+            <strong>Semáforo Executivo Padrão:</strong> mesmo critério para status e riscos em todos os projetos.
+          </li>
+          <li>
+            <strong>Artefatos Essenciais, não Burocracia:</strong> mínimos necessários, padronizados, reutilizáveis.
+          </li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'papeis-macro',
+    badge: '1',
+    label: 'Papéis',
+    title: 'Papéis Macro (RACI de alto nível)',
+    content: (
+      <>
+        <ul>
+          <li>
+            <strong>Sponsor/Comitê Executivo (Diretoria, p.ex. Reginaldo):</strong> aprova visão, orçamento e Gates; remove
+            impedimentos críticos.
+          </li>
+          <li>
+            <strong>Liderança Operacional (p.ex. Raul):</strong> priorização operacional, capacidade, ganhos de eficiência.
+          </li>
+          <li>
+            <strong>PMO (Leonardo):</strong> define método, garante governança, planeja/acompanha portfólio, reporta status e riscos,
+            cobra decisões nos Gates.
+          </li>
+          <li>
+            <strong>Gerente de Projeto (GP):</strong> planejamento detalhado, execução, gestão de riscos/issues, comunicação.
+          </li>
+          <li>
+            <strong>Leads de Área:</strong> Produto (Maxwell), Pedagógico (Erica), Operações (Raul), Marketing, Tecnologia (Douglas e
+            Matheus B.), Financeiro (Angelita), Jurídico (Angelita), Suprimentos (Darlete), Relacionamento/CS (Paula Maria), etc.
+          </li>
+        </ul>
+        <blockquote>Nota: A RACI detalhada por projeto será derivada deste macro, no artefato “RACI do Projeto”.</blockquote>
+      </>
+    ),
+  },
+  {
+    id: 'visao-pipeline',
+    badge: '2',
+    label: 'Pipeline',
+    title: 'Visão Geral do Pipeline',
+    content: (
+      <>
+        <p>
+          <strong>Fases &amp; Gates:</strong>
+        </p>
+        <ul className="link-list">
+          <li>
+            <a
+              href="https://www.notion.so/Fluxo-do-Escrit-rio-de-Projetos-PMO-Educacross-277e9f3bb81e80e5a81bc16554c44efa?pvs=21"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <strong>Intake &amp; Triage → G0 (Aceitar para Descoberta?)</strong>
+            </a>
+          </li>
+          <li>
+            <a
+              href="https://www.notion.so/Fluxo-do-Escrit-rio-de-Projetos-PMO-Educacross-277e9f3bb81e80e5a81bc16554c44efa?pvs=21"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <strong>Descoberta &amp; Iniciação → G1 (Aprovar Charter/Business Case?)</strong>
+            </a>
+          </li>
+          <li>
+            <a
+              href="https://www.notion.so/Fluxo-do-Escrit-rio-de-Projetos-PMO-Educacross-277e9f3bb81e80e5a81bc16554c44efa?pvs=21"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <strong>Planejamento Detalhado → G2 (Plano pronto para Execução?)</strong>
+            </a>
+          </li>
+          <li>
+            <a
+              href="https://www.notion.so/Fluxo-do-Escrit-rio-de-Projetos-PMO-Educacross-277e9f3bb81e80e5a81bc16554c44efa?pvs=21"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <strong>Execução &amp; Monitoramento → G3 (Liberar para Implantação?)</strong>
+            </a>
+          </li>
+          <li>
+            <a
+              href="https://www.notion.so/Fluxo-do-Escrit-rio-de-Projetos-PMO-Educacross-277e9f3bb81e80e5a81bc16554c44efa?pvs=21"
+              target="_blank"
+              rel="noopener noreferrer"
+            >
+              <strong>Lançamento &amp; Estabilização → G4 (Aprovar Encerramento?)</strong>
+            </a>
+          </li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'g0',
+    badge: '3',
+    label: 'G0',
+    title: 'G0: Intake &amp; Triage',
+    content: (
+      <>
+        <p>
+          <strong>Objetivo:</strong> Filtrar e priorizar demandas antes de consumir capacidade do time.
+        </p>
+        <h3>Atividades-chave</h3>
+        <ol>
+          <li>
+            <strong>Canal único de entrada:</strong> formulário padrão Bitrix24 + backlog de iniciativas.
+          </li>
+          <li>
+            <strong>Triagem semanal PMO + Liderança:</strong> classificação em Estratégico / Melhoria / Operacional.
+          </li>
+          <li>
+            <strong>Checklist de prontidão:</strong> objetivos claros, patrocinador, valor esperado.
+          </li>
+        </ol>
+        <h3>Saídas &amp; Critérios para G1</h3>
+        <ul>
+          <li>Iniciativas priorizadas com sponsor designado.</li>
+          <li>Estimativa macro de esforço/benefício.</li>
+          <li>Decisão: segue para Descoberta ou retorna ao backlog.</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'g1',
+    badge: '4',
+    label: 'G1',
+    title: 'G1: Descoberta &amp; Iniciação',
+    content: (
+      <>
+        <p>
+          <strong>Objetivo:</strong> Validar problema/oportunidade, hipóteses e viabilidade antes de grandes investimentos.
+        </p>
+        <h3>Atividades-chave</h3>
+        <ol>
+          <li>
+            <strong>Discovery rápido:</strong> entrevistas com áreas, análise de dados, benchmark.
+          </li>
+          <li>
+            <strong>Canvas de problema/objetivos:</strong> definição de sucesso, hipóteses críticas.
+          </li>
+          <li>
+            <strong>Charter/Business Case preliminar:</strong> objetivos, escopo macro, estimativa de investimento, riscos iniciais.
+          </li>
+          <li>
+            <strong>RACI macro:</strong> sponsor, PM, áreas envolvidas, principais responsáveis.
+          </li>
+          <li>
+            <strong>Kickoff Executivo (opcional):</strong> alinhar expectativas com diretoria.
+          </li>
+        </ol>
+        <h3>Saídas &amp; Critérios para G2</h3>
+        <ul>
+          <li>Charter/Business Case aprovado pela diretoria.</li>
+          <li>Financiamento aprovado (se necessário) + estimativa inicial de ROI.</li>
+          <li>Time núcleo definido (PM + leads de área).
+          </li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'g2',
+    badge: '5',
+    label: 'G2',
+    title: 'G2: Planejamento Detalhado',
+    content: (
+      <>
+        <p>
+          <strong>Objetivo:</strong> Construir plano executável, com escopo detalhado, cronograma, orçamento e estratégia de execução.
+        </p>
+        <h3>Atividades-chave</h3>
+        <ol>
+          <li>
+            <strong>Plano do Projeto:</strong> escopo detalhado, entregáveis, milestones, orçamento, riscos, stakeholders.
+          </li>
+          <li>
+            <strong>Plano de Comunicação:</strong> cadência de ritos, responsáveis, canais, templates.
+          </li>
+          <li>
+            <strong>Plano de Entrega:</strong> detalhamento das ondas de execução, recursos necessários, integrações.
+          </li>
+          <li>
+            <strong>Plano de Implantação (se aplicável):</strong> preparação de operações, treinamento, suporte.
+          </li>
+        </ol>
+        <h3>Saídas &amp; Critérios para G3</h3>
+        <ul>
+          <li>Plano aprovado pelo sponsor + diretoria.</li>
+          <li>Recursos alocados (pessoas, orçamento, tecnologia).</li>
+          <li>Riscos críticos endereçados com planos de resposta.</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'g3',
+    badge: '6',
+    label: 'G3',
+    title: 'G3: Execução &amp; Monitoramento',
+    content: (
+      <>
+        <p>
+          <strong>Objetivo:</strong> Entregar escopo aprovado, monitorar indicadores e assegurar qualidade.
+        </p>
+        <h3>Atividades-chave</h3>
+        <ol>
+          <li>
+            <strong>Ritos de acompanhamento:</strong> dailies/squads, semanal de execução, quinzenal executivo.
+          </li>
+          <li>
+            <strong>Gestão de riscos/issues:</strong> registro, análise, plano de ação, escalonamento quando necessário.
+          </li>
+          <li>
+            <strong>Status Report quinzenal:</strong> semáforo padrão (escopo, prazo, orçamento, riscos, próximos passos).
+          </li>
+          <li>
+            <strong>Gestão de mudanças:</strong> avaliação de impactos, aprovação de sponsor, atualização de plano.
+          </li>
+        </ol>
+        <h3>Saídas &amp; Critérios para G4</h3>
+        <ul>
+          <li>Entregáveis concluídos e validados (critérios de aceite).</li>
+          <li>Plano de lançamento/implantação aprovado.</li>
+          <li>Riscos críticos mitigados ou com plano robusto.</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'g4',
+    badge: '7',
+    label: 'G4',
+    title: 'G4: Lançamento &amp; Estabilização',
+    content: (
+      <>
+        <p>
+          <strong>Objetivo:</strong> Colocar solução em produção/mercado com handover estruturado e operação assistida.
+        </p>
+        <h3>Atividades-chave</h3>
+        <ol>
+          <li>
+            <strong>Plano de Handover:</strong> transição para operações, documentação final, treinamento.
+          </li>
+          <li>
+            <strong>Runbook/Playbook:</strong> procedimentos operacionais, SLAs, contatos de suporte.
+          </li>
+          <li>
+            <strong>Monitoramento pós-lançamento:</strong> KPIs, indicadores de adoção/satisfação, correções rápidas.
+          </li>
+          <li>
+            <strong>Reunião de Encerramento:</strong> validação de entregas, lições aprendidas, assinatura de termo de encerramento.
+          </li>
+        </ol>
+        <h3>Saídas &amp; Critérios para Pós-Projeto</h3>
+        <ul>
+          <li>Operação assumida pela área responsável.</li>
+          <li>KPIs de estabilização acompanhados (30/60/90 dias).</li>
+          <li>Plano de benefícios/ROI iniciado.</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'pos-projeto',
+    badge: '8',
+    label: 'Pós-Projeto',
+    title: 'Pós-Projeto &amp; Benefícios',
+    content: (
+      <>
+        <p>
+          <strong>Objetivo:</strong> Garantir captura de valor, medir resultados e retroalimentar aprendizados.
+        </p>
+        <h3>Atividades-chave</h3>
+        <ol>
+          <li>
+            <strong>Follow-ups 30/60/90:</strong> monitorar KPIs, confirmar geração de valor, ajustar rota.
+          </li>
+          <li>
+            <strong>Relatório de Benefícios:</strong> ROI, ganhos financeiros/operacionais, indicadores qualitativos.
+          </li>
+          <li>
+            <strong>Encerramento formal:</strong> aprovação da diretoria, arquivamento de artefatos, atualização de base de conhecimento.
+          </li>
+        </ol>
+        <h3>Saídas</h3>
+        <ul>
+          <li>Relatório de Benefícios assinado.</li>
+          <li>Atualização do portfólio (pipeline) com status final e aprendizados.</li>
+          <li>Inputs para melhoria contínua de metodologia.</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'governanca-portfolio',
+    badge: '9',
+    label: 'Portfólio',
+    title: 'Governança de Portfólio',
+    content: (
+      <>
+        <ul>
+          <li>
+            <strong>Comitê de Portfólio (mensal):</strong> diretoria + PMO + líderes. Prioriza pipeline, resolve conflitos de capacidade, aprova Gates críticos.
+          </li>
+          <li>
+            <strong>Dashboard Executivo:</strong> visão consolidada (status, riscos, ROI, próximos marcos).
+          </li>
+          <li>
+            <strong>Rituais Operacionais:</strong> semanal/quinzenal com líderes para acompanhamento detalhado.
+          </li>
+          <li>
+            <strong>Gestão de Demandas:</strong> triagem contínua, backlog priorizado, comunicação transparente.
+          </li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'artefatos-por-area',
+    badge: '10',
+    label: 'Artefatos',
+    title: 'Artefatos por Área',
+    content: (
+      <>
+        <div className="flow-artifacts-grid">
+          <article>
+            <h3>PMO</h3>
+            <ul>
+              <li>Playbook do fluxo</li>
+              <li>Modelos de Charter, Plano de Projeto, Status Report</li>
+              <li>Template de Comitê de Portfólio</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Produto</h3>
+            <ul>
+              <li>Backlog priorizado, Discovery Canvas, Plano de Releases</li>
+              <li>Matriz de impactos pedagógicos</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Operações</h3>
+            <ul>
+              <li>Planos de implantação e suporte</li>
+              <li>Rituais de operação assistida</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Comercial/CS</h3>
+            <ul>
+              <li>Planos de comunicação com clientes</li>
+              <li>Scripts e playbooks de atendimento</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Tecnologia</h3>
+            <ul>
+              <li>Roadmap técnico, documentação de integrações</li>
+              <li>Plano de contingência</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Financeiro/Jurídico</h3>
+            <ul>
+              <li>Análises de viabilidade, contratos, compliance</li>
+            </ul>
+          </article>
+        </div>
+      </>
+    ),
+  },
+  {
+    id: 'bitrix',
+    badge: '11',
+    label: 'Bitrix24',
+    title: 'Bitrix24 como Hub Operacional',
+    content: (
+      <>
+        <ul>
+          <li>Pipeline único com etapas por Gate (G0→G4) e automações.</li>
+          <li>Templates de tarefas por etapa, checklists padronizados.</li>
+          <li>Integração com dashboards executivos (Looker/Metabase).</li>
+          <li>Alertas automáticos para riscos, prazos e aprovações pendentes.</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'politicas',
+    badge: '12',
+    label: 'Políticas',
+    title: 'Políticas e Diretrizes',
+    content: (
+      <>
+        <ul>
+          <li>Todo projeto estratégico deve passar por G0 antes de consumir squads/times.</li>
+          <li>Gate sem decisão em 5 dias úteis → escalonamento automático para diretoria.</li>
+          <li>Alterações de escopo acima de 15% exigem revisão de Business Case.</li>
+          <li>Artefatos mínimos obrigatórios para cada Gate (charter, plano, status report, handover, relatório de benefícios).</li>
+        </ul>
+      </>
+    ),
+  },
+  {
+    id: 'implantacao',
+    badge: '13',
+    label: 'Implantação',
+    title: 'Roadmap de Implantação',
+    content: (
+      <>
+        <div className="flow-implementation">
+          <article>
+            <h3>Fase 1 — 0 a 30 dias</h3>
+            <ul>
+              <li>Implantação do fluxo G0→G4 no Bitrix24.</li>
+              <li>Treinamento rápido com líderes de área + Squads.</li>
+              <li>Criação do comitê executivo quinzenal.</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Fase 2 — 30 a 60 dias</h3>
+            <ul>
+              <li>Integração com dashboards Looker/Metabase.</li>
+              <li>Implantação dos artefatos padrão (charter, plano, status report, handover).</li>
+              <li>Ritos de monitoramento consolidados.</li>
+            </ul>
+          </article>
+          <article>
+            <h3>Fase 3 — 60 a 90 dias</h3>
+            <ul>
+              <li>Operação assistida com squads (governança em produção).</li>
+              <li>Dashboards executivos com indicadores de ROI.</li>
+              <li>Retro contínua da metodologia + otimização de automações.</li>
+            </ul>
+          </article>
+        </div>
+      </>
+    ),
+  },
+];
+
+export default flowSections;
